refactor: Separates HttpsSocket class in hpp and cpp files
